% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp20test-package.R
\docType{package}
\name{cpp20test-package}
\alias{cpp20test}
\alias{cpp20test-package}
\title{cpp20test: A Package For Testing cpp20}
\description{
Unit tests for the R package cpp20.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NicChr/cpp20/issues}
}

}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/0000-0002-9743-7342}{ORCID})

}
\keyword{internal}
